// frontend/src/app/models/artikal.ts
export class Artikal{
    idPred: number;
    sifra: number;
    naziv: string;
    jedinicaMere: string;
    stopaPoreza: number;
    proizvodjac: string;
    tip: string;
    zemljaPorekla: string;
    straniNaziv: string;
    barkod: number;
    carinskaTarifa: number;
    ekoTaksa: boolean;
    akcize: boolean;
    minimalneZeljeneZalihe: number;
    maksimalneZeljeneZalihe: number;
    opis: string;
    deklaracija: string;
    slika: string;
    dodeljenaKategorija: boolean;
}